+++ crates/ruff_linter/src/rules/ruff/rules/zip_instead_of_pairwise.rs
